requirements.txt
tensorflow==2.19.0
pandas==2.2.3
scikit-learn==1.6.1
xgboost==3.0.1
numpy
matplotlib==3.10.3
seaborn==0.13.2
streamlit==1.45.1
mlflow==2.22.0
joblib==1.5.0
python-docx==1.1.2
pillow
ipykernel==6.29.5